move _calc_m0_from_spectrum and _f_to_df to FM_utils
