+++ mikeio/dfsu_spectral.py
@@ -9,7 +9,7 @@ from mikecore.DfsuFile import DfsuFile, DfsuFileType
 from .dfsu import _Dfsu
 from .dataset import Dataset, DataArray
 from .dfsutil import _get_item_info, _valid_item_numbers, _valid_timesteps
-from .spatial.FM_utils import _plot_spectrum
+from .spatial.FM_utils import _plot_spectrum, _calc_m0_from_spectrum
 
 
 class DfsuSpectral(_Dfsu):
@@ -338,7 +338,7 @@ class DfsuSpectral(_Dfsu):
             significant wave height values
         """
         if isinstance(spectrum, DataArray):
-            m0 = self._calc_m0_from_spectrum(
+            m0 = _calc_m0_from_spectrum(
                 spectrum.to_numpy(),
                 self.frequencies,
                 self.directions,
@@ -347,40 +347,7 @@ class DfsuSpectral(_Dfsu):
             )
         else:
 
-            m0 = self._calc_m0_from_spectrum(
+            m0 = _calc_m0_from_spectrum(
                 spectrum, self.frequencies, self.directions, tail, m0_only=True
             )
         return 4 * np.sqrt(m0)
-
-    @staticmethod
-    def _calc_m0_from_spectrum(spec, f, dir=None, tail=True, m0_only=False):
-        if f is None:
-            raise ValueError(
-                "Moments cannot be calculated because dfsu has no frequency axis"
-            )
-        df = DfsuSpectral._f_to_df(f)
-
-        if dir is None:
-            ee = spec
-        else:
-            nd = len(dir)
-            dtheta = (dir[-1] - dir[0]) / (nd - 1)
-            ee = np.sum(spec, axis=-2) * dtheta
-
-        m0 = np.dot(ee, df)
-        if tail:
-            m0 = m0 + ee[..., -1] * f[-1] * 0.25
-        return m0
-
-    @staticmethod
-    def _f_to_df(f):
-        """Frequency bins for equidistant or logrithmic frequency axis"""
-        if np.isclose(np.diff(f).min(), np.diff(f).max()):
-            # equidistant frequency bins
-            return (f[1] - f[0]) * np.ones_like(f)
-        else:
-            # logarithmic frequency bins
-            freq_factor = f[1] / f[0]
-            fm1 = np.insert(f, 0, f[0] / freq_factor)
-            fp1 = np.append(f, f[-1] * freq_factor)
-            return 0.5 * (np.diff(fm1) + np.diff(fp1))
